[FSSDK-11544] core restructure
