+++ OptimizelySDK/Config/DatafileProjectConfig.cs
@@ -18,7 +18,6 @@
 #define USE_ODP
 #endif
 
-using System;
 using System.Collections.Generic;
 using System.Linq;
 using Newtonsoft.Json;
